Update sharing-public.md
Removed the notion of "easy" and "easily" - If it were easy, you wouldn't need documentation.

Want to make it easy? Teach all the basics through the cli as soon as they install it.

I also moved the command right up to the top.

Making people wade through words to get the commands is not a best practice.

Don't bury the lead, as it were.

Signed-off-by: Brian Ball <ideabrian@gmail.com>